requirements.txt
openai
requests
beautifulsoup4
pyyaml
ddgs
rich>=13.0.0
click>=8.0.0